+++ classes/view/components/bombDisplay.hpp
@@ -14,7 +14,7 @@ public:
 
     virtual void drawShape() const override;
     virtual void setCenter(const Point &dest) override;
-    virtual void setSize(int newSize) override;
+    virtual void setSize(double newSize) override;
 
     virtual std::string type() const {return ComponentDisplay::type() != " " ? ComponentDisplay::type() + text.getText() : " ";}
 };
