+++ classes/view/shapes/rectangle.hpp
@@ -4,12 +4,12 @@
 #include "shape.hpp"
 
 class Rectangle : public Shape {
-    int w, h;
+    double w, h;
     Fl_Boxtype type;
 public:
-    Rectangle(const Point &center, int fillColour, int size, Fl_Boxtype type) 
+    Rectangle(const Point &center, int fillColour, double size, Fl_Boxtype type) 
                 : Shape{center, fillColour, size}, w{size}, h{size}, type{type} {}
-    Rectangle(const Point &center, int fillColour, int w, int h, Fl_Boxtype type) 
+    Rectangle(const Point &center, int fillColour, double w, double h, Fl_Boxtype type) 
                 : Shape{center, fillColour, w}, w{w}, h{h}, type{type} {}
                                                                                           
     Rectangle(const Rectangle &) = default;
